+++ src/Http/Requests/CreateLink.php
@@ -31,6 +31,7 @@ class CreateLink extends FormRequest implements CreateLinkContract
             'link_type' => 'required|string|exclude_with:link_type_to_create|exists:link_types,slug',
             'link_type_to_create' => 'sometimes|string|min:1|max:255',
             'omnidirectional' => 'sometimes|bool',
+            'link_group_id' => 'sometimes|nullable|required_without:link_type,|numeric|exists:link_groups,id',
             'source_type' => ['sometimes', 'nullable', 'string', Rule::in(CreateLinkForm::$acceptedTypes)],
             'source_id' => 'required|numeric',
             'target_type' => ['sometimes', 'nullable', 'string', Rule::in(CreateLinkForm::$acceptedTypes)],