+++ src/Http/Requests/CreateTaxon.php
@@ -21,9 +21,18 @@ class CreateTaxon extends FormRequest implements CreateTaxonContract
     public function rules()
     {
         return [
-            'name' => 'required|min:2|max:255',
+            'name' => 'required|string|min:2|max:255',
+            'slug' => 'sometimes|nullable|string|max:255',
             'parent_id' => 'nullable|exists:taxons,id',
             'priority' => 'nullable|integer',
+            'ext_title' => 'sometimes|nullable|string|max:511',
+            'meta_description' => 'sometimes|nullable|string|max:4096',
+            'meta_keywords' => 'sometimes|nullable|string|max:1024',
+            'subtitle' => 'sometimes|nullable|string|max:255',
+            'excerpt' => 'sometimes|nullable|string|max:65534',
+            'description' => 'sometimes|nullable|string|max:65534',
+            'top_content' => 'sometimes|nullable|string|max:65534',
+            'bottom_content' => 'sometimes|nullable|string|max:65534',
             'images' => 'nullable',
             'images.*' => 'image|mimes:jpg,jpeg,pjpg,png,gif,webp',
         ];